requirement.txt
streamlit
pandas
tensorflow
scikit-learn
rapidfuzz
Pillow
openpyxl